+++ cres_validation/__init__.py
@@ -1,6 +1,6 @@
 """Package cres_validation - Validation et correction de fichiers CSV pour le projet CRES"""
 
-__version__ = "0.1.0"
+__version__ = "0.2.1"
 
 # Exports publics
 from cres_validation.columns_number_validator import (
@@ -9,6 +9,7 @@ from cres_validation.columns_number_validator import (
     validate_csv,
 )
 from cres_validation.config import get_config
+from cres_validation.convert_to_parquet import convert_csv_to_parquet
 from cres_validation.convert_txt_to_csv import (
     convert_txt_file_to_csv,
     convert_txt_to_csv,
@@ -23,4 +24,5 @@ __all__ = [
     "convert_txt_to_csv",
     "convert_txt_file_to_csv",
     "detect_encoding",
+    "convert_csv_to_parquet",
 ]